requirements.txt
langgraph>=0.1.10
typing_extensions>=4.8.0
google-api-python-client>=2.151.0
google-auth>=2.33.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=1.26.4
